Fix One-Hot Encoding of State

+++ src/lifelonglearning.py
@@ -13,9 +13,9 @@ import torch.nn.functional as F
 class DQN(nn.Module):
     def __init__(self, input_dim, n_actions):
         super(DQN, self).__init__()
-        self.fc1 = nn.Linear(input_dim, 64)
-        self.fc2 = nn.Linear(64, 64)
-        self.fc3 = nn.Linear(64, n_actions)
+        self.fc1 = nn.Linear(input_dim, 64, dtype=torch.float64)
+        self.fc2 = nn.Linear(64, 64, dtype=torch.float64)
+        self.fc3 = nn.Linear(64, n_actions, dtype=torch.float64)
     
     def forward(self, x):
         x = F.relu(self.fc1(x))
